Rename / slice 폴더 생성 후 posts 파일 이동
