+++ src/pages/Create.jsx
@@ -4,7 +4,7 @@ import Header from "../common/Header";
 import Container from "../common/Container";
 import { nanoid } from "@reduxjs/toolkit";
 import { useDispatch, useSelector } from "react-redux";
-import { addPost } from "../redux/posts";
+import { addPost } from "../redux/slice/posts";
 
 export default function Create() {
   const posts = useSelector((state) => state.posts);
